backend/requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-CORS==3.0.10
python-dotenv==0.19.0